requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML models and processing
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
whisper>=1.1.10
speechrecognition>=3.10.0
librosa>=0.10.1
soundfile>=0.12.1
scikit-learn>=1.3.0

# Image and video processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Firebase integration
firebase-admin>=6.4.0

# Report generation
reportlab>=4.0.0
openpyxl>=3.1.0

# System monitoring and optimization
psutil>=5.9.0

# Utilities
pydantic>=2.4.0
aiofiles>=23.2.0